rust: refactor `message` and `sequence` decoder modules
Extracts the `Decoder` and `State` types into their own modules
